RAG+Agent+Youtube_Search/requirements.txt
gradio>=4.0.0
pinecone-client>=3.0.0
sentence-transformers>=2.2.2
groq>=0.3.0
torch>=2.0.0
nltk>=3.8.1
spacy>=3.7.0
youtube-transcript-api>=0.6.1
google-api-python-client>=2.0.0
python-dotenv>=1.0.0 